+++ src/components/sections/summary-section.tsx
@@ -25,17 +25,14 @@ export function SummarySection() {
     experience: (typeof experienceData.items)[0],
     index: number
   ) => (
-    <Card
-      key={`summary-${index}`}
-      className="sm:h-[calc(100vh-30rem)] h-[calc(100vh-27rem)] overflow-y-auto flex-[1_1_0%]" // Adjust height for mobile
-    >
+    <Card key={`summary-${index}`} className="overflow-y-auto">
       <CardHeader>
-        <div className="flex flex-col md:flex-row md:items-center md:justify-between gap-4">
-          <div>
+        <div className="flex flex-row w-full md:items-center md:justify-between gap-4">
+          <div className="w-full">
             <CardTitle>{experience.company}</CardTitle>
             <CardDescription>{experience.position}</CardDescription>
           </div>
-          <Badge variant="outline" className="w-fit">
+          <Badge variant="outline" className="w-fit h-fit">
             {experience.period}
           </Badge>
         </div>
@@ -43,16 +40,8 @@ export function SummarySection() {
       <CardContent className="overflow-y-auto max-h-full px-4">
         {/* Scrollable content */}
         <div className="px-2 space-y-4">
-          <div>
-            <h4 className="font-medium">
-              {experienceData.responsibilitiesLabel}
-            </h4>
-            <p>{experience.shortResponsibilities}</p>
-          </div>
-          <div>
-            <h4 className="font-medium">{experienceData.achievementsLabel}</h4>
-            <p>{experience.shortAchievements}</p>
-          </div>
+          <p className="text-sm mb-0">{experience.shortResponsibilities}</p>
+          <p className="text-sm">{experience.shortAchievements}</p>
         </div>
       </CardContent>
     </Card>
@@ -60,7 +49,7 @@ export function SummarySection() {
 
   // Render brief intro card
   const renderBriefIntroCard = () => (
-    <Card className="flex-[1_1_0]">
+    <Card className="row-span-1 col-span-8 md:row-span-2 md:col-span-6">
       <CardHeader>
         <CardTitle>{summaryData.title}</CardTitle>
       </CardHeader>
@@ -72,7 +61,7 @@ export function SummarySection() {
 
   // Render skills card
   const renderSkillsCard = () => (
-    <Card className="flex-[1_1_0]">
+    <Card className="row-span-2 col-span-4 md:col-span-3 max-h-full">
       <CardHeader>
         <CardTitle>{skillsData.title}</CardTitle>
       </CardHeader>
@@ -90,7 +79,7 @@ export function SummarySection() {
 
   // Render education card
   const renderEducationCard = () => (
-    <Card className="flex-[1_1_0]">
+    <Card className="row-span-1 col-span-8 md:row-span-2 md:col-span-3">
       <CardHeader>
         <CardTitle>{educationData.title}</CardTitle>
       </CardHeader>
@@ -111,20 +100,21 @@ export function SummarySection() {
       id="summary"
       title={summaryData.title}
       description={""}
-      useCarouselOnMobile={true}
+      className="h-full overflow-y-auto"
     >
-      {/* Top row with Brief Intro, Skills, and Education cards */}
-      <div className="flex flex-col lg:flex-row gap-4 mb-4">
-        {renderBriefIntroCard()}
-        {renderSkillsCard()}
-        {renderEducationCard()}
-      </div>
-
-      {/* Render the first three experience cards in a row */}
-      <div className="flex flex-col lg:flex-row gap-4">
-        {experienceData.items
-          .slice(0, 3)
-          .map((summary, index) => renderExperienceCard(summary, index))}
+      <div className="flex flex-col justify-center overflow-x-hidden">
+        {/* Top row with Brief Intro, Skills, and Education cards */}
+        <div className="grid grid-rows-2 grid-cols-12 gap-4 mb-4 ">
+          {renderBriefIntroCard()}
+          {renderSkillsCard()}
+          {renderEducationCard()}
+        </div>
+        {/* Render the first three experience cards in a row */}
+        <div className="grid grid-cols-1 md:grid-cols-3 gap-4">
+          {experienceData.items
+            .slice(0, 3)
+            .map((summary, index) => renderExperienceCard(summary, index))}
+        </div>
       </div>
     </Section>
   );